Use Audio Component of react-article-components
