+++ admin/public/styles/keystone.less
@@ -45,7 +45,7 @@
 
 // REACT FIELDS
 @import "react/react.less";
-@import (inline) "../../../node_modules/react-article-components/lib/main.css";
+@import (inline) "../../../node_modules/react-article-components/src/styles/article.css";
 @import "react/react-slider.less";
