+++ fields/types/html/editor/audio/audio-block.js
@@ -1,38 +1,26 @@
 'use strict';
-
-import { Entity } from 'draft-js';
-import { AudioItem } from '../../../../../admin/client/components/AudioGrid';
+import { Audio } from 'react-article-components';
+import AtomicBlockRendererMixin from '../mixins/atomic-block-renderer-mixin';
 import React from 'react';
 
-export default class AudioBlock extends React.Component {
+export default class AudioBlock extends AtomicBlockRendererMixin(React.Component) {
   constructor(props) {
     super(props);
-    this.audio = this._getValue() || {};
-  }
-
-  componentWillUnmount() {
-      delete this.audio;
-  }
-
-  _getValue() {
-      const entityKey = this.props.block.getEntityAt(0);
-      return entityKey ? Entity.get(entityKey).getData(): null;
   }
 
   render() {
-      let { className } = this.props;
-      if (!this.audio) {
+      if (!this.state.data) {
           return null;
       }
 
       return (
-          <AudioItem
-              audio={_.get(this.audio, 'url')}
-              coverPhoto={_.get(this.audio, 'coverPhoto')}
-              description={_.get(this.audio, 'description')}
-              id={this.audio.id}
-              title={_.get(this.audio, 'title')}
+          <div
+              contentEditable={false}
+              >
+          <Audio
+            {...this.state.data}
           />
+  </div>
       );
   }
 };
