+++ package.json
@@ -80,7 +80,7 @@
     "react-addons-css-transition-group": "0.14.7",
     "react-addons-pure-render-mixin": "^0.14.8",
     "react-alt-text": "1.1.0",
-    "react-article-components": "^0.1.0",
+    "react-article-components": "^0.1.10",
     "react-color": "1.3.7",
     "react-day-picker": "1.3.0",
     "react-dnd": "^2.1.2",
